Add saving account and change int to float

+++ src/main/java/rocks/zipcode/atm/CashMachine.java
@@ -30,7 +30,7 @@ public class CashMachine {
     }
 
     //float
-    public void deposit(int amount) {
+    public void deposit(float amount) {
         if (accountData != null) {
             tryCall(
                     () -> bank.deposit(accountData, amount),
@@ -39,7 +39,7 @@ public class CashMachine {
         }
     }
     //float
-    public void withdraw(int amount) {
+    public void withdraw(float amount) {
         if (accountData != null) {
             tryCall(
                     () -> bank.withdraw(accountData, amount),
